+++ src/main/java/elucent/rootsclassic/RegistryManager.java
@@ -5,13 +5,6 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import elucent.rootsclassic.block.BlockAestheticStandingStone;
-import elucent.rootsclassic.block.BlockAltar;
-import elucent.rootsclassic.block.BlockBrazier;
-import elucent.rootsclassic.block.BlockFlareOrchid;
-import elucent.rootsclassic.block.BlockImbuer;
-import elucent.rootsclassic.block.BlockMidnightBloom;
-import elucent.rootsclassic.block.BlockMortar;
-import elucent.rootsclassic.block.BlockRadiantDaisy;
 import elucent.rootsclassic.block.BlockStandingStoneAccelerator;
 import elucent.rootsclassic.block.BlockStandingStoneEntangler;
 import elucent.rootsclassic.block.BlockStandingStoneGrower;
@@ -21,11 +14,27 @@ import elucent.rootsclassic.block.BlockStandingStoneRepulsor;
 import elucent.rootsclassic.block.BlockStandingStoneT1;
 import elucent.rootsclassic.block.BlockStandingStoneT2;
 import elucent.rootsclassic.block.BlockStandingStoneVacuum;
+import elucent.rootsclassic.block.altar.BlockAltar;
+import elucent.rootsclassic.block.altar.TileEntityAltar;
+import elucent.rootsclassic.block.altar.TileEntityAltarRenderer;
+import elucent.rootsclassic.block.brazier.BlockBrazier;
+import elucent.rootsclassic.block.brazier.TileEntityBrazier;
+import elucent.rootsclassic.block.brazier.TileEntityBrazierRenderer;
+import elucent.rootsclassic.block.flowers.BlockFlareOrchid;
+import elucent.rootsclassic.block.flowers.BlockMidnightBloom;
+import elucent.rootsclassic.block.flowers.BlockRadiantDaisy;
+import elucent.rootsclassic.block.imbuer.BlockImbuer;
+import elucent.rootsclassic.block.imbuer.TileEntityImbuer;
+import elucent.rootsclassic.block.imbuer.TileEntityImbuerRenderer;
+import elucent.rootsclassic.block.mortar.BlockMortar;
+import elucent.rootsclassic.block.mortar.TileEntityMortar;
+import elucent.rootsclassic.block.mortar.TileEntityMortarRenderer;
 import elucent.rootsclassic.component.ComponentManager;
 import elucent.rootsclassic.entity.EntityAccelerator;
 import elucent.rootsclassic.entity.EntityTileAccelerator;
 import elucent.rootsclassic.entity.skeleton.EntityFrozenKnight;
 import elucent.rootsclassic.entity.skeleton.RenderFrozenKnight;
+import elucent.rootsclassic.item.FuelManager;
 import elucent.rootsclassic.item.ItemCrystalStaff;
 import elucent.rootsclassic.item.ItemDragonsEye;
 import elucent.rootsclassic.item.ItemDruidArmor;
@@ -49,15 +58,6 @@ import elucent.rootsclassic.item.ItemRunicFocus;
 import elucent.rootsclassic.item.ItemStaff;
 import elucent.rootsclassic.ritual.RitualManager;
 import elucent.rootsclassic.tileentity.TileEntityAestheticStandingStone;
-import elucent.rootsclassic.tileentity.TileEntityAltar;
-import elucent.rootsclassic.tileentity.TileEntityAltarRenderer;
-import elucent.rootsclassic.tileentity.TileEntityBrazier;
-import elucent.rootsclassic.tileentity.TileEntityBrazierRenderer;
-import elucent.rootsclassic.tileentity.TileEntityDruidChalice;
-import elucent.rootsclassic.tileentity.TileEntityImbuer;
-import elucent.rootsclassic.tileentity.TileEntityImbuerRenderer;
-import elucent.rootsclassic.tileentity.TileEntityMortar;
-import elucent.rootsclassic.tileentity.TileEntityMortarRenderer;
 import elucent.rootsclassic.tileentity.TileEntityStandingStoneAccelerator;
 import elucent.rootsclassic.tileentity.TileEntityStandingStoneEntangler;
 import elucent.rootsclassic.tileentity.TileEntityStandingStoneGrower;
@@ -252,7 +252,6 @@ public class RegistryManager {
     GameRegistry.registerTileEntity(TileEntityMortar.class, "te_mortar");
     GameRegistry.registerTileEntity(TileEntityImbuer.class, "te_imbuer");
     GameRegistry.registerTileEntity(TileEntityAltar.class, "te_altar");
-    GameRegistry.registerTileEntity(TileEntityDruidChalice.class, "te_chalice");
     GameRegistry.registerTileEntity(TileEntityBrazier.class, "te_brazier");
     GameRegistry.registerTileEntity(TileEntityStandingStoneVacuum.class, "te_vacuum");
     GameRegistry.registerTileEntity(TileEntityStandingStoneRepulsor.class, "te_repulsor");
