+++ src/main/java/elucent/rootsclassic/block/altar/BlockAltar.java
@@ -1,21 +1,22 @@
-package elucent.rootsclassic.block;
+package elucent.rootsclassic.block.altar;
 
 import elucent.rootsclassic.Roots;
-import elucent.rootsclassic.tileentity.TileEntityMortar;
+import elucent.rootsclassic.block.TEBlockBase;
 import net.minecraft.block.ITileEntityProvider;
 import net.minecraft.block.material.Material;
 import net.minecraft.block.state.IBlockState;
 import net.minecraft.tileentity.TileEntity;
+import net.minecraft.util.BlockRenderLayer;
 import net.minecraft.util.math.AxisAlignedBB;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.IBlockAccess;
 import net.minecraft.world.World;
 
-public class BlockMortar extends TEBlockBase implements ITileEntityProvider {
+public class BlockAltar extends TEBlockBase implements ITileEntityProvider {
 
-  private static final AxisAlignedBB AXIS_ALIGNED_BB = new AxisAlignedBB(0.3125, 0, 0.3125, 0.6875, 0.3125, 0.6875);
+  private static final AxisAlignedBB AXIS_ALIGNED_BB = new AxisAlignedBB(0.0, 0, 0.0, 1.0, 0.75, 1.0);
 
-  public BlockMortar() {
+  public BlockAltar() {
     super(Material.GROUND);
     setCreativeTab(Roots.tab);
     setHardness(1.0f);
@@ -37,7 +38,12 @@ public class BlockMortar extends TEBlockBase implements ITileEntityProvider {
   }
 
   @Override
+  public boolean canRenderInLayer(IBlockState state, BlockRenderLayer layer) {
+    return (layer == BlockRenderLayer.CUTOUT);
+  }
+
+  @Override
   public TileEntity createNewTileEntity(World worldIn, int meta) {
-    return new TileEntityMortar();
+    return new TileEntityAltar();
   }
 }
