+++ src/main/java/elucent/rootsclassic/block/imbuer/TileEntityImbuerRenderer.java
@@ -1,4 +1,4 @@
-package elucent.rootsclassic.tileentity;
+package elucent.rootsclassic.block.imbuer;
 
 import org.lwjgl.opengl.GL11;
 import net.minecraft.client.Minecraft;
