+++ src/main/java/elucent/rootsclassic/block/mortar/TileEntityMortar.java
@@ -1,10 +1,11 @@
-package elucent.rootsclassic.tileentity;
+package elucent.rootsclassic.block.mortar;
 
 import java.util.ArrayList;
 import java.util.List;
 import elucent.rootsclassic.RegistryManager;
 import elucent.rootsclassic.component.ComponentManager;
 import elucent.rootsclassic.component.ComponentRecipe;
+import elucent.rootsclassic.tileentity.TEBase;
 import net.minecraft.block.state.IBlockState;
 import net.minecraft.client.resources.I18n;
 import net.minecraft.entity.item.EntityItem;
