+++ src/main/java/elucent/rootsclassic/ritual/RitualManager.java
@@ -3,6 +3,8 @@ package elucent.rootsclassic.ritual;
 import java.util.ArrayList;
 import elucent.rootsclassic.RegistryManager;
 import elucent.rootsclassic.Util;
+import elucent.rootsclassic.block.altar.TileEntityAltar;
+import elucent.rootsclassic.block.brazier.TileEntityBrazier;
 import elucent.rootsclassic.ritual.rituals.RitualCauseRain;
 import elucent.rootsclassic.ritual.rituals.RitualCrafting;
 import elucent.rootsclassic.ritual.rituals.RitualEngravedSword;
@@ -14,8 +16,6 @@ import elucent.rootsclassic.ritual.rituals.RitualMassBreed;
 import elucent.rootsclassic.ritual.rituals.RitualSacrifice;
 import elucent.rootsclassic.ritual.rituals.RitualSummoning;
 import elucent.rootsclassic.ritual.rituals.RitualTimeShift;
-import elucent.rootsclassic.tileentity.TileEntityAltar;
-import elucent.rootsclassic.tileentity.TileEntityBrazier;
 import net.minecraft.entity.monster.EntityCaveSpider;
 import net.minecraft.entity.monster.EntityCreeper;
 import net.minecraft.entity.monster.EntityEnderman;
